refactor(move): Extract error module from utils package into project source root. Update imports where used.
